% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as.matrix.wbw::WhiteboxRaster}
\alias{as.matrix.wbw::WhiteboxRaster}
\title{Return matrix from WhiteboxRaster}
\usage{
\method{as.matrix}{`wbw::WhiteboxRaster`}(x, ...)
}
\arguments{
\item{x}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
matrix
}
\description{
Return matrix from WhiteboxRaster
}
\keyword{transform}
