% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as.vector.wbw::WhiteboxRaster}
\alias{as.vector.wbw::WhiteboxRaster}
\title{Return vector from WhiteboxRaster}
\usage{
\method{as.vector}{`wbw::WhiteboxRaster`}(x, ...)
}
\arguments{
\item{x}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
vector
}
\description{
Return vector from WhiteboxRaster
}
\keyword{transform}
