% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_wbw_raster}
\alias{as_wbw_raster}
\title{Convert SpatRaster to WhiteboxRaster}
\usage{
as_wbw_raster(x)
}
\arguments{
\item{x}{SpatRaster object}
}
\value{
\link{WhiteboxRaster} object
}
\description{
Convert SpatRaster to WhiteboxRaster
}
\examples{
\dontshow{if (requireNamespace("terra", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(terra)
f <- system.file("extdata/dem.tif", package = "wbw")
rast(f) |>
  as_wbw_raster()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=WhiteboxRaster]{WhiteboxRaster()}}
}
\keyword{Converts}
\keyword{SpatRaster}
\keyword{[WhiteboxRaster]}
\keyword{a}
\keyword{object}
\keyword{terra}
\keyword{to}
