% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_package}
\alias{check_package}
\alias{check_env}
\alias{check_input_file}
\alias{skip_if_no_wbw}
\title{Check if package is installed}
\usage{
check_package(package)

check_env(env = wbe)

check_input_file(file_name, type)

skip_if_no_wbw()
}
\description{
Check if package is installed

Check if whitebox environment is present

Check input file extension

Skip tests if we don't have the 'wbw' module
}
\keyword{internal}
