% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{ncell}
\alias{ncell}
\alias{nrow}
\alias{ncol}
\title{Dimensions of a WhiteboxRaster or WhiteboxVector objects}
\usage{
ncell(x)

nrow(x)

ncol(x)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\value{
\code{integer}
}
\description{
Dimensions of a WhiteboxRaster or WhiteboxVector objects
}
\keyword{utils}
