% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_documentation.R
\name{expect_snapshot}
\alias{expect_snapshot}
\title{Tinytest Snapshot
https://github.com/etiennebacher/astgrepr/
blob/ea91137bdb10d22c7a988a8cb1b0bc896935fb0d/R/tinytest.R}
\usage{
expect_snapshot(label, current)
}
\description{
Tinytest Snapshot
https://github.com/etiennebacher/astgrepr/
blob/ea91137bdb10d22c7a988a8cb1b0bc896935fb0d/R/tinytest.R
}
\keyword{internal}
