% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as_matrix}
\alias{as_matrix}
\title{Return matrix from WhiteboxRaster}
\usage{
as_matrix(x, raw = FALSE)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{raw}{logical. Should the raw data be returned (\code{raw = TRUE}) or
should \code{NoData} values be transformed to \code{NA} (\code{raw = FALSE})?}
}
\value{
matrix
}
\description{
Return matrix from WhiteboxRaster
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  as_matrix(raw = TRUE)
}
}
\keyword{transform}
