% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{max.wbw::WhiteboxRaster}
\alias{max.wbw::WhiteboxRaster}
\title{Get maximum value from WhiteboxRaster}
\usage{
\method{max}{`wbw::WhiteboxRaster`}(object, ...)
}
\arguments{
\item{object}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
numeric value
}
\description{
Get maximum value from WhiteboxRaster
}
