% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{mean.wbw::WhiteboxRaster}
\alias{mean.wbw::WhiteboxRaster}
\title{Get mean value from WhiteboxRaster}
\usage{
\method{mean}{`wbw::WhiteboxRaster`}(x, ...)
}
\arguments{
\item{x}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
numeric value
}
\description{
Get mean value from WhiteboxRaster
}
