% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{min.wbw::WhiteboxRaster}
\alias{min.wbw::WhiteboxRaster}
\title{Get minimum value from WhiteboxRaster}
\usage{
\method{min}{`wbw::WhiteboxRaster`}(object, ...)
}
\arguments{
\item{object}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
numeric value
}
\description{
Get minimum value from WhiteboxRaster
}
