% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.wbw::WbwRaster}
\alias{print.wbw::WbwRaster}
\title{Print method for WbwRaster}
\usage{
\method{print}{`wbw::WbwRaster`}(x, ...)
}
\arguments{
\item{x}{\link{WbwRaster} object to print}

\item{...}{additional arguments passed to print}
}
\description{
Print method for WbwRaster
}
