% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_geotiff_tags}
\alias{print_geotiff_tags}
\title{Print GeoTIFFs tags}
\usage{
print_geotiff_tags(file_name)
}
\arguments{
\item{file_name}{\code{character}, path to raster file.}
}
\description{
This tool can be used to view the tags contained within a GeoTiff file.
Viewing the tags of a GeoTiff file can be useful when trying to import
the GeoTiff to different software environments. The user must specify the
name of a GeoTiff file and the tag information will be printed in the
console. Note that tags that contain greater than 100
values will be truncated in the output. GeoKeys will also be interpreted
as per the GeoTIFF specification.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
print_geotiff_tags(raster_path)
}

}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#print_geotiff_tags}
}
