% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{res}
\alias{res}
\alias{xres}
\alias{yres}
\title{Get WhiteboxRaster resolution}
\usage{
res(x)

xres(x)

yres(x)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\value{
\code{double}
}
\description{
Get WhiteboxRaster resolution
}
\keyword{utils}
