% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{stdev}
\alias{stdev}
\title{Get Standard Deviation value from WhiteboxRaster}
\usage{
stdev(x)
}
\arguments{
\item{x}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
numeric value
}
\description{
Get Standard Deviation value from WhiteboxRaster
}
