% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary.wbw::WhiteboxRaster}
\alias{summary.wbw::WhiteboxRaster}
\alias{max.wbw::WhiteboxRaster}
\alias{mean.wbw::WhiteboxRaster}
\alias{median.wbw::WhiteboxRaster}
\alias{min.wbw::WhiteboxRaster}
\alias{stdev}
\alias{variance}
\title{Summary Statistics for WhiteboxRaster}
\usage{
\method{summary}{`wbw::WhiteboxRaster`}(object, ...)

\method{max}{`wbw::WhiteboxRaster`}(object, ...)

\method{mean}{`wbw::WhiteboxRaster`}(x, ...)

\method{median}{`wbw::WhiteboxRaster`}(x, na.rm = FALSE, ...)

\method{min}{`wbw::WhiteboxRaster`}(object, ...)

stdev(x)

variance(x)
}
\arguments{
\item{object}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}

\item{x}{WhiteboxRaster object}

\item{na.rm}{logical indicating whether NA values should
be stripped (not used)}
}
\value{
numeric value
}
\description{
Computes summary statistics for cells in a \link{WhiteboxRaster} object.
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  summary()
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  max()
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  mean()
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  median()
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  min()
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  stdev()
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  variance()
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#raster_summary_stats}
}
\keyword{stats}
