% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.wbw::WhiteboxRaster}
\alias{summary.wbw::WhiteboxRaster}
\title{Summary method for WhiteboxRaster}
\usage{
\method{summary}{`wbw::WhiteboxRaster`}(object, ...)
}
\arguments{
\item{object}{\link{WhiteboxRaster} object}

\item{...}{additional arguments passed to summary}
}
\description{
Summary method for WhiteboxRaster
}
