% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wbw_adaptive_filter}
\alias{wbw_adaptive_filter}
\title{Adaptive Filter}
\usage{
wbw_adaptive_filter(x, filter_size_x = 11L, filter_size_y = 11L, threshold = 2)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}

\item{threshold}{\code{double}}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
Applies an adaptive filter to reduce random noise (shot noise) in a raster
image. The filter modifies pixel values only where they differ substantially
from their neighbors.
}
\details{
The algorithm calculates the average value in a moving window centered on
each grid cell. If the absolute difference between the window mean and the
center cell value exceeds the user-defined threshold, the output cell is
assigned the mean value. Otherwise, it retains its original value.

Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should be odd,
positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_adaptive_filter(filter_size_x = 3L, filter_size_y = 3L)
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#adaptive_filter}
}
\keyword{image_processing}
