% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorphometry.R
\name{wbw_aspect}
\alias{wbw_aspect}
\title{Aspect}
\usage{
wbw_aspect(dem, z_factor = 1)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{z_factor}{\code{double}, Z conversion factor is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z conversion factor}
}
\value{
\link{WhiteboxRaster} object containing aspect in degrees
}
\description{
Calculates slope aspect (i.e., slope orientation in degrees clockwise from
north) for each grid cell in an input DEM.
}
\details{
For DEMs in projected coordinate systems, the tool uses the 3rd-order
bivariate Taylor polynomial method described by Florinsky (2016). Based on a
polynomial fit of elevations within the 5x5 neighborhood surrounding each
cell, this method is more robust against outlier elevations (noise) than
other methods.

For DEMs in geographic coordinate systems (i.e., angular units), the tool
uses the 3x3 polynomial fitting method for equal angle grids also described
by Florinsky (2016).
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_aspect()
}
}
\references{
Gallant, J. C., and J. P. Wilson, 2000, Primary topographic attributes,
in Terrain Analysis: Principles and Applications, edited by J. P. Wilson
and J. C. Gallant pp. 51-86, John Wiley, Hoboken, N.J.

Florinsky, I. (2016). Digital terrain analysis in soil science and
geology. Academic Press.

For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#aspect}
}
\seealso{
\code{\link[=wbw_to_degrees]{wbw_to_degrees()}}, \code{\link[=wbw_to_radians]{wbw_to_radians()}}, \code{\link[=wbw_slope]{wbw_slope()}}
}
\keyword{geomorphometry}
