% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{wbw_bilateral_filter}
\alias{wbw_bilateral_filter}
\title{Bilateral Filter}
\usage{
wbw_bilateral_filter(x, sigma_dist = 0.75, sigma_int = 1)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{sigma_dist}{\code{double}, standard deviation distance parameter in
\strong{pixels}.}

\item{sigma_int}{\code{double}, standard deviation intensity parameter,
in the same units as z-units of input raster \code{x} (usually, meters).}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
Applies an edge-preserving smoothing filter that reduces noise while
preserving important edges in the image.
}
\details{
Bilateral filtering is a non-linear technique introduced by Tomasi and
Manduchi (1998). The filter combines spatial and intensity domains to
preserve edges while smoothing. Unlike the Gaussian filter, the bilateral
filter weights pixels based on both their spatial distance and intensity
similarity to the center pixel.

The size of the filter is determined by setting the standard deviation
distance parameter (\code{sigma_dist}); the larger the standard deviation the
larger the resulting filter kernel.
The standard deviation can be any number in the range
0.5-20 and is specified in the unit of pixels. The standard deviation
intensity parameter (\code{sigma_int}), specified in the same units as the
z-values, determines the intensity domain contribution to kernel weightings.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_bilateral_filter(sigma_dist = 1.5, sigma_int = 1.1)
}
}
\references{
Tomasi, C., & Manduchi, R. (1998, January). Bilateral filtering for gray
and color images. In null (p. 839). IEEE.

For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#bilateral_filter}
}
\seealso{
\code{\link[=wbw_mean_filter]{wbw_mean_filter()}}, \code{\link[=wbw_gaussian_filter]{wbw_gaussian_filter()}}
}
\keyword{image_processing}
