% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wbw_conservative_smoothing_filter}
\alias{wbw_conservative_smoothing_filter}
\title{Conservative Smoothing Filter}
\usage{
wbw_conservative_smoothing_filter(x, filter_size_x = 3L, filter_size_y = 3L)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
This tool performs a conservative smoothing filter on a raster image.
A conservative smoothing filter can be used to remove short-range
variability in an image, effectively acting to smooth the image.
It is particularly useful for eliminating local spikes and reducing the
noise in an image.
}
\details{
The algorithm operates by calculating the minimum and maximum neighbouring
values surrounding a grid cell. If the cell at the centre of the
kernel is greater than the calculated maximum value, it is replaced
with the maximum value in the output image. Similarly, if the cell
value at the kernel centre is less than the neighbouring minimum value,
the corresponding grid cell in the output image is replaced with the
minimum value.

Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should be odd,
positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_conservative_smoothing_filter(filter_size_x = 3L, filter_size_y = 3L)
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#conservative_smoothing_filter}
}
\seealso{
\code{\link[=wbw_bilateral_filter]{wbw_bilateral_filter()}}, \code{\link[=wbw_gaussian_filter]{wbw_gaussian_filter()}},
\code{\link[=wbw_mean_filter]{wbw_mean_filter()}}
}
\keyword{image_processing}
