% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{wbw_download_sample_data}
\alias{wbw_download_sample_data}
\title{Download Sample Data}
\usage{
wbw_download_sample_data(data_set = NULL, path = NULL)
}
\arguments{
\item{data_set}{\code{character}, dataset name. See Details}

\item{path}{\code{character}, path to where download sample datasets. If
\code{NULL}, the currect working directory is used}
}
\description{
There are a number of available sample datasets that can be readily used
to test Whitebox Workflows for Python.
}
\details{
Available datasets:\tabular{lll}{
   \strong{Dataset Name} \tab \strong{Description} \tab \strong{Size} \cr
   Guelph_landsat \tab Landsat 5 sub-area (7 bands) \tab 10.9 MB \cr
   Grand_Junction \tab Small DEM in high-relief terrain \tab 5.8 MB \cr
   GTA_lidar \tab Airborne lidar point cloud (LAZ) \tab 54.3 MB \cr
   jay_brook \tab Airborne lidar point cloud (LAZ) \tab 76.3 MB \cr
   Jay_State_Forest \tab Lidar-derived DEM \tab 27.7 MB \cr
   Kitchener_lidar \tab Airborne lidar point cloud (LAZ) \tab 41.6 MB \cr
   London_air_photo \tab High-resolution RGB air photo \tab 87.3 MB \cr
   mill_brook \tab Airborne lidar point cloud (LAZ) \tab 49.9 MB \cr
   peterborough_drumlins \tab Lidar-derived DEM \tab 22.0 MB \cr
   Southern_Ontario_roads \tab Vector roads layer \tab 7.1 MB \cr
   StElisAk \tab Airborne lidar point cloud (LAZ) \tab 54.5 MB \cr
}
}
\examples{
\dontrun{
# Download sample data
wbw_download_sample_data(data_set = "Guelph_landsat", path = tempdir())
}
}
\keyword{system}
