% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{wbw_ext}
\alias{wbw_ext}
\title{Get WhiteboxExtent}
\usage{
wbw_ext(x)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\description{
Get WhiteboxExtent
}
