% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{wbw_gaussian_curvature}
\alias{wbw_gaussian_curvature}
\title{Gaussian Curvature}
\usage{
wbw_gaussian_curvature(dem, log_transform = FALSE, z_factor = 1)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{log_transform}{\code{logical}, default \code{FALSE}. Wheter
log-transform the output raster or not. See details.}

\item{z_factor}{\code{double}, Z conversion factor is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z conversion factor}
}
\value{
\link{WhiteboxRaster} object in units of \eqn{m^{-2}}.
}
\description{
This tool calculates the Gaussian curvature from a digital elevation
model (\eqn{dem}). Gaussian curvature is the product of maximal and
minimal curvatures, and retains values in each point of the topographic
surface after its bending without breaking, stretching, and
compressing (Florinsky, 2017).

Gaussian curvature is measured in units of \eqn{m^{-2}}.
}
\details{
Curvature values are often very small and as such the user may opt to
log-transform the output raster (\eqn{log_transform}). Transforming
the values applies the equation by Shary et al. (2002):

\deqn{Θ' = sign(Θ) ln(1 + 10^n|Θ|)}

where \eqn{Θ} is the parameter value and \eqn{n} is dependent on the
grid cell size.

For DEMs in projected coordinate systems, the tool uses the
3rd-order bivariate Taylor polynomial method described by
Florinsky (2016). Based on a polynomial fit of the elevations
within the 5x5 neighbourhood surrounding each cell, this method
is considered more robust against outlier elevations (noise)
than other methods.

For DEMs in geographic coordinate systems (i.e. angular units), the
tool uses the 3x3 polynomial fitting method for equal angle grids also
described by Florinsky (2016).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_gaussian_curvature()
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#gaussian_curvature}

Florinsky, I. (2016). Digital terrain analysis in soil science and
geology. Academic Press. \if{html}{\out{<br>}}
Florinsky, I. V. (2017). An illustrated introduction to general
geomorphometry. Progress in Physical Geography, 41(6), 723-752. \if{html}{\out{<br>}}
Shary P. A., Sharaya L. S. and Mitusov A. V. (2002) Fundamental
quantitative methods of land surface analysis. Geoderma 107: 1–32. \if{html}{\out{<br>}}
}
\seealso{
\code{\link[=wbw_maximal_curvature]{wbw_maximal_curvature()}}, \code{\link[=wbw_minimal_curvature]{wbw_minimal_curvature()}}
}
\keyword{geomorphometry}
