% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{wbw_gaussian_filter}
\alias{wbw_gaussian_filter}
\title{Gaussian filter}
\usage{
wbw_gaussian_filter(x, sigma = 0.75)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{sigma}{\code{double}, standard deviation distance parameter in
\strong{units of grid cells}. Should be in the range 0.5-20.}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
This tool can be used to perform a Gaussian filter on a raster image.
A Gaussian filter can be used to emphasize the longer-range variability in
an image, effectively acting to smooth the image. This can be useful for
reducing the noise in an image.
}
\details{
The algorithm operates by convolving a kernel of weights with each grid cell
and its neighbours in an image. The weights of the convolution kernel are
determined by the 2-dimensional Gaussian (i.e. normal) curve, which gives
stronger weighting to cells nearer the kernel centre. It is this
characteristic that makes the Gaussian filter an attractive alternative for
image smoothing and noise reduction than the \link{wbw_mean_filter}.
The size of the filter is determined by setting the
standard deviation parameter (\code{sigma}), which is in units of grid cells;
the larger the standard deviation the larger the resulting filter kernel.
The standard deviation can be any number in the range 0.5-20.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_gaussian_filter(sigma = 1.5)
}
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#gaussian_filter}
}
\seealso{
\code{\link[=wbw_mean_filter]{wbw_mean_filter()}}
}
\keyword{image_processing}
