% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{wbw_get_name}
\alias{wbw_get_name}
\title{Get File Name from Path}
\usage{
wbw_get_name(path)
}
\arguments{
\item{path}{\code{character}, file path}
}
\value{
\code{character}, file name
}
\description{
Extracts the file name from a full path string.
}
\keyword{internal}
