% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wbw_high_pass_median_filter}
\alias{wbw_high_pass_median_filter}
\title{High Pass Median Filter}
\usage{
wbw_high_pass_median_filter(
  x,
  filter_size_x = 11L,
  filter_size_y = 11L,
  sig_digits = 2L
)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}

\item{sig_digits}{\code{integer}}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
This tool performs a high-pass median filter on a raster image.
High-pass filters can be used to emphasize the short-range variability in
an image. The algorithm operates essentially by subtracting the value at
the grid cell at the centre of the window from the median value in the
surrounding neighbourhood (i.e. window.)
}
\details{
Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should
be odd, positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_high_pass_median_filter(filter_size_x = 3L, filter_size_y = 3L)
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#high_pass_median_filter}
}
\seealso{
\code{\link[=wbw_median_filter]{wbw_median_filter()}}, \code{\link[=wbw_high_pass_filter]{wbw_high_pass_filter()}}
}
\keyword{image_processing}
