% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{wbw_hillshade}
\alias{wbw_hillshade}
\title{Hillshade}
\usage{
wbw_hillshade(dem, azimuth = 315, altitude = 30, z_factor = 1)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{azimuth}{\code{double}, illumination source azimuth or
sun direction (0 to 360 degrees)}

\item{altitude}{\code{double}, the altitude of the illumination sources.
i.e. the elevation of the sun above the horizon, measured as an angle from
0 to 90 degrees}

\item{z_factor}{\code{double}, Z conversion factor is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z conversion factor}
}
\value{
\link{WhiteboxRaster} object
}
\description{
This tool performs a hillshade operation (also called shaded relief) on an
input digital elevation model (DEM).
}
\details{
The hillshade value (HS) of a DEM grid cell is calculate as:
\deqn{HS = \frac{\tan(s)}{\sqrt{1 - \tan(s)^2}} \times
[\frac{\sin(Alt)}{\tan(s)} - \cos(Alt) \times \sin(Az - a)]}
where \eqn{s} and \eqn{a} are the local slope gradient and aspect
(orientation) respectively and \eqn{Alt} and \eqn{Az} are the illumination
source altitude and azimuth respectively. Slope and aspect are calculated
using Horn's (1981) 3rd-order finate difference method.

If the DEM is in the geographic coordinate system (latitude and longitude),
the following equation is used:
\deqn{zfactor = \frac{1.0}{111320.0 \times \cos(midlat)}}

where \eqn{midlat} is the latitude of the centre of the raster,
in radians.
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_hillshade()
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#hillshade}

Gallant, J. C., and J. P. Wilson, 2000, Primary topographic attributes,
in Terrain Analysis: Principles and Applications, edited by J. P. Wilson
and J. C. Gallant pp. 51-86, John Wiley, Hoboken, N.J.
}
\seealso{
\code{\link[=wbw_multidirectional_hillshade]{wbw_multidirectional_hillshade()}}
}
\keyword{geomorphometry}
