% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{wbw_max_procs}
\alias{wbw_max_procs}
\title{Set Maximum Parallel Processors}
\usage{
wbw_max_procs(max_procs = -1)
}
\arguments{
\item{max_procs}{\code{integer} Number of processors to use. Use -1 for all
available processors, or a positive integer to limit processor usage.}
}
\description{
Determines the number of processors used by functions that are parallelized.
If set to -1 (\code{max_procs=-1}), the default, all available processors will be
used. To limit processing, set \code{max_procs} to a positive whole number less
than the number of system processors.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
x <- wbw_read_raster(raster_path)

# Use 1 processor
wbw_max_procs(1)
system.time(wbw_slope(x))

# Use all available processors
wbw_max_procs(-1)
system.time(wbw_slope(x))
}
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#max_procs}
}
\keyword{system}
