% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{wbw_mean_curvature}
\alias{wbw_mean_curvature}
\title{Mean Curvature}
\usage{
wbw_mean_curvature(dem, log_transform = FALSE, z_factor = 1)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{log_transform}{\code{logical}, default \code{FALSE}. Wheter
log-transform the output raster or not. See details.}

\item{z_factor}{\code{double}, Z conversion factor is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z conversion factor}
}
\value{
\link{WhiteboxRaster} object.
}
\description{
This tool calculates the mean curvature from a digital elevation
model (\eqn{dem}).

WhiteboxTools reports curvature in radians multiplied by 100 for easier
interpretation because curvature values are typically very small.
}
\details{
Curvature values are often very small and as such the user may opt to
log-transform the output raster (\eqn{log_transform}). Transforming
the values applies the equation by Shary et al. (2002):

\deqn{Θ' = sign(Θ) ln(1 + 10^n|Θ|)}

where \eqn{Θ} is the parameter value and \eqn{n} is dependent on the
grid cell size.

For DEMs in projected coordinate systems, the tool uses the
3rd-order bivariate Taylor polynomial method described by
Florinsky (2016). Based on a polynomial fit of the elevations
within the 5x5 neighbourhood surrounding each cell, this method
is considered more robust against outlier elevations (noise)
than other methods.

For DEMs in geographic coordinate systems (i.e. angular units), the
tool uses the 3x3 polynomial fitting method for equal angle grids also
described by Florinsky (2016).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_mean_curvature()
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#mean_curvature}

Florinsky, I. (2016). Digital terrain analysis in soil science and
geology. Academic Press. \if{html}{\out{<br>}}
Florinsky, I. V. (2017). An illustrated introduction to general
geomorphometry. Progress in Physical Geography, 41(6), 723-752. \if{html}{\out{<br>}}
Shary P. A., Sharaya L. S. and Mitusov A. V. (2002) Fundamental
quantitative methods of land surface analysis. Geoderma 107: 1–32. \if{html}{\out{<br>}}
}
\seealso{
\code{\link[=wbw_gaussian_curvature]{wbw_gaussian_curvature()}}, \code{\link[=wbw_maximal_curvature]{wbw_maximal_curvature()}}
}
\keyword{geomorphometry}
