% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wbw_minimum_filter}
\alias{wbw_minimum_filter}
\title{Minimum Filter}
\usage{
wbw_minimum_filter(x, filter_size_x = 11L, filter_size_y = 11L)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
Assigns each cell in the output grid the minimum value in a moving window
centred on each grid cell in the input raster.
}
\details{
Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should
be odd, positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_minimum_filter(filter_size_x = 3L, filter_size_y = 3L)
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#minimum_filter}
}
\seealso{
\code{\link[=wbw_mean_filter]{wbw_mean_filter()}}, \code{\link[=wbw_high_pass_filter]{wbw_high_pass_filter()}},
\code{\link[=wbw_high_pass_median_filter]{wbw_high_pass_median_filter()}}
}
\keyword{image_processing}
