% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{wbw_multidirectional_hillshade}
\alias{wbw_multidirectional_hillshade}
\title{Multidirectional Hillshade}
\usage{
wbw_multidirectional_hillshade(
  dem,
  altitude = 30,
  z_factor = 1,
  full_360_mode = FALSE
)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{altitude}{\code{double}, the altitude of the illumination sources.
i.e. the elevation of the sun above the horizon, measured as an angle from
0 to 90 degrees}

\item{z_factor}{\code{double}, Z conversion factor is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z conversion factor}

\item{full_360_mode}{\code{boolean}, default \code{FALSE}. I.e. whether or
not to use full 360-degrees of illumination sources. When \code{FALSE}
(default) the tool will perform a weighted summation of the hillshade
images from four illumination azimuth positions at 225, 270, 315, and
360 (0) degrees, given weights of 0.1, 0.4, 0.4, and 0.1 respectively. When
run in the full 360-degree mode, eight illumination source azimuths are
used to calculate the output at 0, 45, 90, 135, 180, 225, 270, and 315
degrees, with weights of 0.15, 0.125, 0.1, 0.05, 0.1, 0.125, 0.15,
and 0.2 respectively.}
}
\value{
\link{WhiteboxRaster} object
}
\description{
This tool performs a hillshade operation (also called shaded relief) on
an input digital elevation model (DEM) with multiple sources of
illumination.
}
\details{
The hillshade value (HS) of a DEM grid cell is calculate as:
\deqn{HS = \frac{\tan(s)}{\sqrt{1 - \tan(s)^2}} \times [\frac{\sin(Alt)}{\tan(s)} -
  \cos(Alt) \times \sin(Az - a)]}
where \eqn{s} and \eqn{a} are the local slope gradient and aspect (orientation)
respectively and \eqn{Alt} and \eqn{Az} are the illumination source altitude and azimuth
respectively. Slope and aspect are calculated using Horn's (1981) 3rd-order
finate difference method.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_multidirectional_hillshade()
}
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#multidirectional_hillshade}

Horn B.K.P., 1981, Hill shading and the reflectance map, Proceedings of
the I.E.E.E. 69, 14
}
\keyword{geomorphometry}
