% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wbw_olympic_filter}
\alias{wbw_olympic_filter}
\title{Olympic Filter}
\usage{
wbw_olympic_filter(x, filter_size_x = 11L, filter_size_y = 11L)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
This filter is a modification of the \link{wbw_mean_filter}, whereby
the highest and lowest values in the kernel are dropped, and the remaining
values are averaged to replace the central pixel. The result is a
low-pass smoothing filter that is more robust than the \link{wbw_mean_filter},
which is more strongly impacted by the presence of outlier values.
It is named after a system of scoring Olympic events.
}
\details{
Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should
be odd, positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_olympic_filter(filter_size_x = 3L, filter_size_y = 3L)
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#olympic_filter}
}
\seealso{
\code{\link[=wbw_mean_filter]{wbw_mean_filter()}}
}
\keyword{image_processing}
