% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{wbw_plan_curvature}
\alias{wbw_plan_curvature}
\title{Plan Curvature}
\usage{
wbw_plan_curvature(dem, log_transform = FALSE, z_factor = 1)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{log_transform}{\code{logical}, default \code{FALSE}. Wheter
log-transform the output raster or not. See details.}

\item{z_factor}{\code{double}, Z conversion factor is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z conversion factor}
}
\value{
\link{WhiteboxRaster} object.
}
\description{
This tool calculates the plan curvature (i.e. contour curvature), or
the rate of change in aspect along a contour line, from a digital
elevation model (\eqn{dem}). Curvature is the second derivative of the
topographic surface defined by a DEM. Plan curvature characterizes the
degree of flow convergence or divergence within the
landscape (Gallant and Wilson, 2000).
}
\details{
WhiteboxTools reports curvature in degrees multiplied by 100 for easier
interpretation. The Z conversion factor (\eqn{z_factor}) is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z Conversion Factor.

If the DEM is in the geographic coordinate system (latitude and longitude),
the following equation is used:

\deqn{zfactor = \frac{1.0}{111320.0 \times \cos(midlat)}}

The algorithm uses the same formula for the calculation of plan curvature
as Gallant and Wilson (2000). Plan curvature is negative for diverging
flow along ridges and positive for convergent areas,
e.g. along valley bottoms.
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_plan_curvature()
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#plan_curvature}

Gallant, J. C., and J. P. Wilson, 2000, Primary topographic attributes,
in Terrain Analysis: Principles and Applications, edited by J. P.
Wilson and J. C. Gallant pp. 51-86, John Wiley, Hoboken, N.J. \if{html}{\out{<br>}}
}
\seealso{
\code{\link[=wbw_profile_curvature]{wbw_profile_curvature()}}
}
\keyword{geomorphometry}
