% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{wbw_random_sample}
\alias{wbw_random_sample}
\title{Random sample}
\usage{
wbw_random_sample(x, num_samples = 1000L)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{num_samples}{\code{integer}, number of random samples. Should not
exceed total number of cells.}
}
\value{
\link{WhiteboxRaster} object
}
\description{
This tool can be used to create a random sample of grid cells. The user
specifies the \link{WhiteboxRaster}, which is used to determine the grid
dimensions and georeference information for the output \link{WhiteboxRaster}, and
the number of sample random samples (\code{num_samples}).

The output grid will contain \code{num_samples} non-zero grid cells,
randomly distributed throughout the raster grid, and a background
value of \strong{zero}. Every grid cell will have a value from \code{1}
to \code{num_samples}.
}
\details{
This tool is useful when performing statistical analyses on raster images
when you wish to obtain a random sample of data.

Only valid, non-nodata, cells in the base raster will be sampled.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  random_sample(num_samples = 100)
}
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#random_sample}
}
\keyword{math}
