% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wbw_range_filter}
\alias{wbw_range_filter}
\title{Range Filter}
\usage{
wbw_range_filter(x, filter_size_x = 11L, filter_size_y = 11L)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
A range filter assigns to each cell in the output grid the range
(maximum - minimum) of the values contained within a moving window
centred on each grid cell.
}
\details{
Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should
be odd, positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_range_filter(filter_size_x = 3L, filter_size_y = 3L)
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#range_filter}
}
\seealso{
\code{\link[=wbw_minimum_filter]{wbw_minimum_filter()}}, \code{\link[=wbw_maximum_filter]{wbw_maximum_filter()}}
}
\keyword{image_processing}
