% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbw_read.R
\name{wbw_read_raster}
\alias{wbw_read_raster}
\title{Returns a new Raster object, read into memory from a path-file string.}
\usage{
wbw_read_raster(file_name)
}
\description{
Returns a new Raster object, read into memory from a path-file string.
}
