% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbw_read.R
\name{wbw_read_vector}
\alias{wbw_read_vector}
\title{Reads a vector from disc into an in-memory Vector object.}
\usage{
wbw_read_vector(file_name)
}
\description{
Reads a vector from disc into an in-memory Vector object.
}
