% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbw_geomorphometry.R
\name{wbw_ruggedness_index}
\alias{wbw_ruggedness_index}
\title{Terrain Ruggedness Index (TRI)}
\usage{
wbw_ruggedness_index(dem)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\value{
\link{WhiteboxRaster} object containing TRI values
}
\description{
The terrain ruggedness index (TRI) is a measure of local topographic relief.
The TRI calculates the root-mean-square-deviation (RMSD) for each grid cell
in a digital elevation model (DEM), calculating the residuals (i.e.
elevation differences) between a grid cell and its eight neighbours.
}
\details{
Notice that, unlike the output of this tool, the original Riley et al.
(1999) TRI did not normalize for the number of cells in the local window
(i.e. it is a root-square-deviation only). However, using the mean has
the advantage of allowing for the varying number of neighbouring cells
along the grid edges and in areas bordering NoData cells. This modification
does however imply that the output of this tool cannot be directly compared
with the index ranges of level to extremely rugged terrain provided in
Riley et al. (1999)
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_ruggedness_index()
}
}
\references{
Riley, S. J., DeGloria, S. D., and Elliot, R. (1999). Index that quantifies
topographic heterogeneity. Intermountain Journal of Sciences, 5(1-4), 23-27.

For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#ruggedness_index}
}
