% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbw_geomorphometry.R
\name{wbw_slope}
\alias{wbw_slope}
\title{Slope}
\usage{
wbw_slope(dem, units = "degrees", z_factor = 1)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{units}{\code{character},
units of slope: "radians", "degrees", or "percent"}

\item{z_factor}{\code{double}, Z conversion factor is only important
when the vertical and horizontal units are not the same in the DEM.
When this is the case, the algorithm will multiply each elevation in the
DEM by the Z conversion factor}
}
\value{
\link{WhiteboxRaster} object containing slope values
}
\description{
This tool calculates slope gradient (i.e. slope steepness in degrees,
radians, or percent) for each grid cell in an input digital elevation
model (DEM).
}
\details{
The tool uses Horn's (1981) 3rd-order finite difference method to
estimate slope. Given the following clock-type grid cell numbering scheme
(Gallant and Wilson, 2000).
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_slope()
}
}
\references{
Gallant, J. C., and J. P. Wilson, 2000, Primary topographic attributes,
in Terrain Analysis: Principles and Applications, edited by J. P. Wilson
and J. C. Gallant pp. 51-86, John Wiley, Hoboken, N.J.

For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#slope}
}
