% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{wbw_version}
\alias{wbw_version}
\title{\code{wbw_version()}: Gets the \verb{Whitebox Workflows} version}
\usage{
wbw_version()
}
\value{
character. Version Number.
}
\description{
\code{wbw_version()}: Gets the \verb{Whitebox Workflows} version
}
