% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{wbw_write_raster}
\alias{wbw_write_raster}
\title{Writes an in-memory WhiteboxRaster object to file.}
\usage{
wbw_write_raster(x, file_name, compress = TRUE)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{file_name}{\code{character} Path to output file}

\item{compress}{\code{logical} Whether to compress the output file}
}
\description{
Writes an in-memory WhiteboxRaster object to a file in a supported raster
format.
}
\details{
Supported raster formats:
\itemize{
\item GeoTIFF (*.tif, *.tiff)
\item Big GeoTIFF (*.tif, *.tiff)
\item SAGA Binary (*.sdat, *.sgrd)
\item Idrisi (*.rst, *.rdc)
\item Surfer (*.grd)
\item ESRI Binary (*.flt)
\item ESRI BIL (*.bil)
}

The tool can read GeoTIFFs compressed using PackBits, DEFLATE, or LZW
methods.

When writing GeoTIFFs, use \code{compress=TRUE} to enable DEFLATE compression.
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#write_raster}
}
\keyword{io}
