% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WbwRaster.R
\name{WbwRaster}
\alias{WbwRaster}
\title{WbwRaster Class}
\usage{
WbwRaster(name = character(0), source = NULL)
}
\arguments{
\item{name}{Character, name of the raster}

\item{source}{Any, source data for the raster}
}
\description{
WbwRaster Class
}
