% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteboxClasses.R
\name{WhiteboxExtent}
\alias{WhiteboxExtent}
\title{WhiteboxExtent Class}
\usage{
WhiteboxExtent(
  west = numeric(0),
  east = numeric(0),
  south = numeric(0),
  north = numeric(0)
)
}
\arguments{
\item{west}{\code{double} Western boundary coordinate}

\item{east}{\code{double} Eastern boundary coordinate}

\item{south}{\code{double} Southern boundary coordinate}

\item{north}{\code{double} Northern boundary coordinate}
}
\description{
Defines the spatial extent of a raster dataset using coordinates for
the west, east, south, and north boundaries.
}
\keyword{class}
