% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteboxClasses.R
\name{WhiteboxRaster}
\alias{WhiteboxRaster}
\title{WhiteboxRaster Class}
\usage{
WhiteboxRaster(name = character(0), source = NULL)
}
\arguments{
\item{name}{Character, name of the raster}

\item{source}{Any, source data for the raster}
}
\description{
WhiteboxRaster Class
}
\section{Properties}{

\describe{
\item{stats}{Character, summary statistics for the raster}
\item{min}{Numeric, minimum value for the raster}
\item{max}{Numeric, maximum value for the raster}
\item{extent}{WhiteboxExtent, spatial extent of the raster}
}
}

\keyword{A}
\keyword{Python}
\keyword{The}
\keyword{Whitebox}
\keyword{a}
\keyword{access}
\keyword{and}
\keyword{class}
\keyword{dataset}
\keyword{format.}
\keyword{in}
\keyword{including}
\keyword{min/max}
\keyword{name,}
\keyword{pointer,}
\keyword{properties}
\keyword{provides}
\keyword{raster}
\keyword{representing}
\keyword{statistics,}
\keyword{summary}
\keyword{to}
\keyword{values.}
