% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{xy.coords.wbw::WhiteboxRaster}
\alias{xy.coords.wbw::WhiteboxRaster}
\title{Convert WhiteboxRaster to x-y coordinates}
\usage{
`xy.coords.wbw::WhiteboxRaster`(x, ...)
}
\arguments{
\item{x}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
A list with NULL x and y components
}
\description{
This is an internal method used by the plotting system.
It returns NULL coordinates to prevent default plotting behavior.
}
\keyword{internal}
