% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_rast}
\alias{as_rast}
\title{Convert WhiteboxRaster to SpatRaster}
\usage{
as_rast(x)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\value{
SpatRaster object
}
\description{
Convert WhiteboxRaster to SpatRaster
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_slope(units = "r") |>
  as_rast()
}
}
\seealso{
\link{WhiteboxRaster}
}
\keyword{Converts}
\keyword{SpatRaster}
\keyword{[WhiteboxRaster]}
\keyword{object}
\keyword{terra}
\keyword{to}
