% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_wbw_raster}
\alias{as_wbw_raster}
\title{Convert SpatRaster to WhiteboxRaster}
\usage{
as_wbw_raster(x)
}
\arguments{
\item{x}{SpatRaster object}
}
\value{
\link{WhiteboxRaster} object
}
\description{
Convert SpatRaster to WhiteboxRaster
}
\examples{
\dontrun{
library(terra)
raster_path <- system.file("extdata/dem.tif", package = "wbw")
rast(raster_path) |>
  as_wbw_raster()
}
}
\seealso{
\code{\link[=WhiteboxRaster]{WhiteboxRaster()}}
}
\keyword{Converts}
\keyword{SpatRaster}
\keyword{[WhiteboxRaster]}
\keyword{object}
\keyword{terra}
\keyword{to}
