% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_package}
\alias{check_package}
\alias{check_env}
\alias{check_input_file}
\title{Check if package is installed}
\usage{
check_package(package)

check_env(env = wbe)

check_input_file(file_name, type)
}
\description{
Check if package is installed

Check if whitebox environment is present

Check input file extension
}
\keyword{internal}
