% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{num_cells}
\alias{num_cells}
\alias{wbw_rows}
\alias{wbw_cols}
\title{Get dimensions of a WhiteboxRaster or WhiteboxVector object}
\usage{
num_cells(x)

wbw_rows(x)

wbw_cols(x)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\value{
\code{integer} Number of cells in the raster
}
\description{
Get dimensions of a WhiteboxRaster or WhiteboxVector object
}
\keyword{utils}
