% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert WhiteboxRaster to Matrix}
\usage{
as_matrix(x, raw = FALSE)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{raw}{logical. Should the raw data be returned (\code{raw = TRUE}) or should
NoData values be transformed to \code{NA} (\code{raw = FALSE})?}
}
\value{
matrix containing raster values
}
\description{
Converts a WhiteboxRaster object to a matrix. The output maintains the same
dimensions and cell values as the input raster.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  as_matrix(raw = TRUE)
}
}
\keyword{transform}
