% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_geotiff_tags}
\alias{print_geotiff_tags}
\title{Print GeoTIFF Tags}
\usage{
print_geotiff_tags(file_name)
}
\arguments{
\item{file_name}{\code{character}, path to raster file}
}
\description{
Displays the tags contained within a GeoTIFF file. This is useful when
importing GeoTIFF files into different software environments. The tool prints
tag information to the console. Tags containing more than 100 values are
truncated in the output. GeoKeys are interpreted according to the GeoTIFF
specification.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
print_geotiff_tags(raster_path)
}

}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#print_geotiff_tags}
}
\keyword{utils}
