% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{wbw_res}
\alias{wbw_res}
\alias{wbw_xres}
\alias{wbw_yres}
\title{Get WhiteboxRaster resolution}
\usage{
wbw_res(x)

wbw_xres(x)

wbw_yres(x)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\value{
\code{double}
}
\description{
Get WhiteboxRaster resolution
}
\keyword{utils}
