% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{variance}
\alias{variance}
\title{Get Variance value from WhiteboxRaster}
\usage{
variance(x)
}
\arguments{
\item{x}{WhiteboxRaster object}

\item{...}{additional arguments (not used)}
}
\value{
numeric value
}
\description{
Get Variance value from WhiteboxRaster
}
