% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as_vector}
\alias{as_vector}
\title{Convert WhiteboxRaster to Vector}
\usage{
as_vector(x, raw = FALSE)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{raw}{logical. Should the raw data be returned (\code{raw = TRUE}) or should
NoData values be transformed to \code{NA} (\code{raw = FALSE})?}
}
\value{
vector containing raster values
}
\description{
Converts a WhiteboxRaster object to a vector. The output contains all raster
values in row-major order.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  as_vector()
}
}
\keyword{transform}
