% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as_vector}
\alias{as_vector}
\alias{as_vector,WhiteboxRaster-method}
\alias{as_vector,WhiteboxExtent-method}
\title{Convert objects to vectors}
\usage{
\S4method{as_vector}{WhiteboxRaster}(x, raw = FALSE)
\S4method{as_vector}{WhiteboxExtent}(x)
}
\arguments{
\item{x}{Object to convert to vector. Can be:
\itemize{
\item A \link{WhiteboxRaster} object
\item A \link{WhiteboxExtent} object
}}

\item{raw}{logical. For \link{WhiteboxRaster} only: Should the raw data be
returned (\code{raw = TRUE}) or should NoData values be transformed
to \code{NA} (\code{raw = FALSE})?}
}
\value{
A vector, with type depending on the input:
\itemize{
\item For \link{WhiteboxRaster}: vector containing raster values
\item For \link{WhiteboxExtent}: named vector containing extent values
}
}
\description{
Converts various Whitebox objects to vectors:
\itemize{
\item For \link{WhiteboxRaster}: converts raster values to a vector in row-major
order
\item For \link{WhiteboxExtent}: converts extent boundaries to a named vector
}
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
x <- wbw_read_raster(f)

# Return WhiteboxRaster's data:
head(as_vector(x))

# Return WhiteboxExtent's data:
as_vector(wbw_ext(x))

}
\keyword{transform}
