% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{wbw_adaptive_filter}
\alias{wbw_adaptive_filter}
\title{Adaptive filter}
\usage{
wbw_adaptive_filter(x, filter_size_x = 11L, filter_size_y = 11L, threshold = 2)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}

\item{threshold}{\code{double}}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
This tool calculates slope gradient (i.e. slope steepness in degrees,
radians, or percent) for each grid cell in an input digital elevation
model (DEM).
}
\details{
This tool performs a type of adaptive filter on a raster image. An adaptive
filter can be used to reduce the level of random noise (shot noise) in an
image. The algorithm operates by calculating the average value in a moving
window centred on each grid cell. If the absolute difference between the
window mean value and the centre grid cell value is beyond a user-defined
threshold (\code{threshold}), the grid cell in the output image is assigned the
mean value, otherwise it is equivalent to the original value. Therefore,
the algorithm only modifies the image where grid cell values are
substantially different than their neighbouring values.

Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should be odd,
positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_adaptive_filter(filter_size_x = 3L, filter_size_y = 3L)
}
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#adaptive_filter}
}
\keyword{image_processing}
