% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{wbw_bilateral_filter}
\alias{wbw_bilateral_filter}
\title{Bilateral filter}
\usage{
wbw_bilateral_filter(x, sigma_dist = 0.75, sigma_int = 1)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{sigma_dist}{\code{double}, standard deviation distance parameter in
\strong{pixels}.}

\item{sigma_int}{\code{double}, standard deviation intensity parameter,
in the same units as z-units of input raster \code{x} (usually, meters).}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
This tool can be used to perform an edge-preserving smoothing filter, or
bilateral filter, on an image. A bilateral filter can be used to emphasize
the longer-range variability in an image, effectively acting to smooth the
image, while reducing the edge blurring effect common with other types of
smoothing filters. As such, this filter is very useful for reducing the
noise in an image.
}
\details{
Bilateral filtering is a non-linear filtering technique introduced by Tomasi
and Manduchi (1998). The algorithm operates by convolving a kernel of
weights with each grid cell and its neighbours in an image. The bilateral
filter is related to Gaussian smoothing, in that the weights of the
convolution kernel are partly determined by the 2-dimensional Gaussian
(i.e. normal) curve, which gives stronger weighting to cells nearer the
kernel centre. Unlike the \link{wbw_gaussian_filter}, however, the bilateral
kernel weightings are also affected by their similarity to the
intensity value of the central pixel.
Pixels that are very different in intensity from the
central pixel are weighted less, also based on a Gaussian weight
distribution. Therefore, this non-linear convolution filter is determined
by the spatial and intensity domains of a localized pixel neighborhood.

The heavier weighting given to nearer and similar-valued pixels makes the
bilateral filter an attractive alternative for image smoothing and noise
reduction compared to the much-used \link{wbw_mean_filter}.
The size of the filter is determined by setting the standard deviation
distance parameter (\code{sigma_dist}); the larger the standard deviation the
larger the resulting filter kernel.
The standard deviation can be any number in the range
0.5-20 and is specified in the unit of pixels. The standard deviation
intensity parameter (\code{sigma_int}), specified in the same units as the
z-values, determines the intensity domain contribution to kernel weightings.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_bilateral_filter(sigma_dist = 1.5, sigma_int = 1.1)
}
}
\references{
Tomasi, C., & Manduchi, R. (1998, January). Bilateral filtering for gray
and color images. In null (p. 839). IEEE.

For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#bilateral_filter}
}
\seealso{
\code{\link[=wbw_mean_filter]{wbw_mean_filter()}}, \code{\link[=wbw_gaussian_filter]{wbw_gaussian_filter()}}
}
\keyword{image_processing}
