% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wbw_gaussian_filter}
\alias{wbw_gaussian_filter}
\title{Gaussian Filter}
\usage{
wbw_gaussian_filter(x, sigma = 0.75)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{sigma}{\code{double}, standard deviation distance parameter in
\strong{units of grid cells}. Should be in the range 0.5-20.}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
Applies a Gaussian smoothing filter to reduce noise while better preserving
image features compared to a simple mean filter.
}
\details{
The filter applies a 2D Gaussian kernel that weights pixels based on their
distance from the center. This gradual weighting makes it more effective for
noise reduction than the mean filter. The filter size is controlled by the
sigma parameter (0.5-20 grid cells).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_gaussian_filter(sigma = 1.5)
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#gaussian_filter}
}
\seealso{
\code{\link[=wbw_mean_filter]{wbw_mean_filter()}}
}
\keyword{image_processing}
