% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbw_install.R
\name{wbw_install}
\alias{wbw_install}
\title{Install Required Python Modules}
\usage{
wbw_install(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{reticulate::py_install()}}
}
\value{
\code{NULL}, or \code{try-error} (invisibly) on R code execution error.
}
\description{
This function installs the latest \code{numpy}, \code{whitebox-workflows}.
The default uses \code{pip} for package installation.
}
\details{
This function provides a basic wrapper around
\code{reticulate::py_install()}, except it defaults to using the Python package
manager \code{pip} and virtual environment. It creates the \code{r-wbw} virtual
environment in the default location (run \code{reticulate::virtualenv_root()} to
find it) and installs the required python packages.
}
\examples{
\dontrun{

wbw_install()
}
}
