% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{wbw_install}
\alias{wbw_install}
\title{Install Required Python Modules}
\usage{
wbw_install(system = FALSE, force = FALSE, ...)
}
\arguments{
\item{system}{\code{boolean}, Use a \code{system()} call to
\code{python -m pip install --user ...}
instead of \code{reticulate::py_install()}. Default: \code{FALSE}.}

\item{force}{\code{boolean}, Force update (uninstall/reinstall) and ignore
existing installed packages? Default: \code{FALSE}.
Applies to \code{system=TRUE}.}

\item{...}{Additional arguments passed to \code{reticulate::py_install()}}
}
\value{
\code{NULL}, or \code{try-error} (invisibly) on R code execution error.
}
\description{
Install Required Python Modules
}
\details{
This function provides a basic wrapper around
\code{reticulate::py_install()}, except it defaults to using the Python package
manager \code{pip} and virtual environment. It creates the \code{r-wbw}
virtual environment in the default location
(run \code{reticulate::virtualenv_root()} to find it) and installs the
required python packages.
}
\keyword{The}
\keyword{This}
\keyword{\code{numpy},}
\keyword{\code{pip}}
\keyword{\code{whitebox-workflows}.}
\keyword{default}
\keyword{for}
\keyword{function}
\keyword{installation.}
\keyword{installs}
\keyword{latest}
\keyword{package}
\keyword{system}
\keyword{the}
\keyword{uses}
