% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{wbw_mean_filter}
\alias{wbw_mean_filter}
\title{Mean Filter}
\usage{
wbw_mean_filter(x, filter_size_x = 11L, filter_size_y = 11L)
}
\arguments{
\item{x}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{filter_size_x}{\code{integer}, X dimension of the neighbourhood size}

\item{filter_size_y}{\code{integer}, Y dimension of the neighbourhood size}
}
\value{
\link{WhiteboxRaster} object containing filtered values
}
\description{
Applies a mean filter (low-pass filter) to smooth an image by emphasizing
longer-range variability and reducing noise.
}
\details{
Uses an efficient integral image approach (Crow, 1984) that is independent
of filter size. While commonly used, mean filters can be more aggressive in
their smoothing compared to edge-preserving alternatives like the bilateral
filter or Gaussian filter.

Neighbourhood size, or filter size, is specified in the x and y dimensions
using \code{filter_size_x} and \code{filter_size_y} These dimensions should be odd,
positive integer values (e.g. 3L, 5L, 7L, 9L, etc.).
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path) |>
  wbw_mean_filter(filter_size_x = 3L, filter_size_y = 3L)
}
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#mean_filter}
}
\seealso{
\code{\link[=wbw_bilateral_filter]{wbw_bilateral_filter()}}, \code{\link[=wbw_gaussian_filter]{wbw_gaussian_filter()}}
}
\keyword{image_processing}
