% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{wbw_read_raster}
\alias{wbw_read_raster}
\title{Returns a new WhiteboxRaster object, read into memory from a
path-file string.}
\usage{
wbw_read_raster(file_name)
}
\arguments{
\item{file_name}{\code{character}, path to raster file.}
}
\value{
\link{WhiteboxRaster} object
}
\description{
Returns a new WhiteboxRaster object, read into memory from a
path-file string.
}
\examples{
\dontrun{
raster_path <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(raster_path)
}
}
