% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{wbw_read_vector}
\alias{wbw_read_vector}
\title{Reads a vector from disc into an in-memory WbwVector object.}
\usage{
wbw_read_vector(file_name)
}
\arguments{
\item{file_name}{\code{character}, path to ESRI shapefile.}
}
\description{
Reads a vector from disc into an in-memory WbwVector object.
}
