% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomorphometry.R
\name{wbw_ruggedness_index}
\alias{wbw_ruggedness_index}
\title{Terrain Ruggedness Index (TRI)}
\usage{
wbw_ruggedness_index(dem)
}
\arguments{
\item{dem}{Raster object of class \link{WhiteboxRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}
}
\value{
\link{WhiteboxRaster} object containing TRI values
}
\description{
Calculates a measure of local topographic relief. The TRI computes the
root-mean-square-deviation (RMSD) for each grid cell in a DEM by calculating
the residuals between a cell and its eight neighbors.
}
\details{
Unlike the original Riley et al. (1999) TRI, this implementation normalizes
for the number of cells in the local window. This modification allows for
varying numbers of neighboring cells along grid edges and areas bordering
NoData cells. Note that this means output values cannot be directly compared
with the index ranges (level to extremely rugged terrain) provided in Riley
et al. (1999).
}
\examples{
f <- system.file("extdata/dem.tif", package = "wbw")
wbw_read_raster(f) |>
  wbw_ruggedness_index()
}
\references{
Riley, S. J., DeGloria, S. D., and Elliot, R. (1999). Index that quantifies
topographic heterogeneity. Intermountain Journal of Sciences, 5(1-4), 23-27.

For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#ruggedness_index}
}
\keyword{geomorphometry}
