% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbw_write.R
\name{wbw_write_raster}
\alias{wbw_write_raster}
\title{Writes an in-memory Raster object to file.}
\usage{
wbw_write_raster(raster, file_name, compress = TRUE)
}
\arguments{
\item{raster}{Raster object of class \link{WbwRaster}. See \code{\link[=wbw_read_raster]{wbw_read_raster()}} for more details.}

\item{file_name}{\code{character}, path to output file}

\item{compress}{\code{logical}, whether to compress the output file}
}
\description{
Writes an in-memory Raster object to file.
}
\references{
For more information, see \url{https://www.whiteboxgeo.com/manual/wbw-user-manual/book/tool_help.html#write_raster}
}
