% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteboxClasses.R
\name{WhiteboxExtent}
\alias{WhiteboxExtent}
\title{WhiteboxExtent Class}
\usage{
WhiteboxExtent(
  west = numeric(0),
  east = numeric(0),
  south = numeric(0),
  north = numeric(0)
)
}
\arguments{
\item{west}{\code{double}}

\item{east}{\code{double}}

\item{south}{\code{double}}

\item{north}{\code{double}}
}
\description{
WhiteboxExtent Class
}
\keyword{class}
