% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhiteboxRaster.R
\name{WhiteboxRaster}
\alias{WhiteboxRaster}
\title{WhiteboxRaster Class}
\usage{
WhiteboxRaster(name = character(0), source = NULL)
}
\arguments{
\item{name}{Character, name of the raster}

\item{source}{Any, source data for the raster}
}
\description{
WhiteboxRaster Class
}
